Cache

XbrlXML.Cache — Module

Provides interface to local store of files used for parsing XBRL.

XbrlXML.Cache.HttpCache — Type

HttpCache(cache_dir="./cache/", headers=Dict())

Create a cache to store files locally for reuse.

headers are passed to http Downloads.download. Services such as SEC require you to disclose information about your application.

Example

julia> cache = HttpCache("/Users/user/cache/")
/Users/user/cache/

julia> cacheheader!(cache, "User-Agent" => "You youremail@domain.com")
Dict{AbstractString, AbstractString} with 1 entry:
  "User-Agent" => "You youremail@domain.com"

XbrlXML.Cache.cache_edgar_enclosure — Method

cache_edgar_enclosure(cache::HttpCache, enclosure_url)

XbrlXML.Cache.cachedir — Method

cachedir(cache::HttpCache)::String

Return the local directory of a cache.

XbrlXML.Cache.cachefile — Method

cachefile(cache::HttpCache, file_url)::String

Save a file located at file_url to a local cache.

XbrlXML.Cache.cacheheader! — Method

cacheheader!(cache::HttpCache, header::Pair)::Dict

Add a header pair to a cache and return the headers.

XbrlXML.Cache.cacheheaders! — Method

cacheheaders!(cache::HttpCache, header::Vector{Pair})::Dict

Add multiple header pairs to a cache and return the headers.

XbrlXML.Cache.cacheheaders — Method

cacheheaders(cache::HttpCache)::Dict

Return the headers of a cache.

XbrlXML.Cache.purgefile — Method

purgefile(cache::HttpCache, file_url)::Bool

Remove a file, based on its URL, from a local cache.

XbrlXML.Cache.urltopath — Method

urltopath(cache::HttpCache, url)::String

Convert a file's url to a local cache file.